Split ACP tests by mode
