+++ tests/functional/acp_base.php
@@ -1,0 +1,41 @@
+<?php
+/**
+ *
+ * Advertisement management. An extension for the phpBB Forum Software package.
+ *
+ * @copyright (c) 2017 phpBB Limited <https://www.phpbb.com>
+ * @license GNU General Public License, version 2 (GPL-2.0)
+ *
+ */
+
+namespace phpbb\ads\tests\functional;
+
+/**
+* @group functional
+*/
+class acp_base extends \phpbb_functional_test_case
+{
+	/**
+	* {@inheritDoc}
+	*/
+	static protected function setup_extensions()
+	{
+		return array('phpbb/ads');
+	}
+
+	/**
+	* {@inheritDoc}
+	*/
+	public function setUp()
+	{
+		parent::setUp();
+
+		$this->add_lang_ext('phpbb/ads', array(
+			'info_acp_phpbb_ads',
+			'acp',
+		));
+
+		$this->login();
+		$this->admin_login();
+	}
+}
